Support call-methods and reorganize the "scope" of methoddef!
Formerly methoddef! tried to get all methods affiliated with a single
expression. This reworks it to get a single method, but one which
fits the name in the "opening" 1-arg :method expression. This allows
for more robust handling of inner methods but also allows more flexible
handling of methods known only by their signature, such as call-methods.

This adds a utility, methoddefs!, to get multiple methods. This is
primarily intended to facilitate testing.