Move backedge analysis here from Revise, add docs, and release 1.0 (#30)
* Move Revise's BackEdges here as CodeEdges

This package seems like a more natural home. It might be useful in
contexts outside of Revise. In Revise they are a means to an end,
and that's resulted in limited direct testing.

Moving them here also encourages fleshing out the API.
This adds many new utitilies and `show` methods.
Testing is done by checking the selectivity of execution, an easier
but thorough way of examining whether its making the correct decisions.

* Add Documenter docs and bump to version 1.0

* Add some precompiles

* Fix an apparent bug in bodymethod