Exclude typedefs, if desired
By default, LoweredCodeUtils will mark the type-definition lines as part
of the constructor dependencies. When selectively executed, this leads
to attempted redefinition of the type, which generally (but not always)
succeeds if the redefined type is the same as the original.
Exceptions include key types in Core.Compiler and LinearAlgebra.
Hence, it's better to support a mode in which you avoid any type redefinition.

This also:

- switches to using :global as markers for the beginning and end of a
typedef
- lowers the optimization level for compilation for reduced latency
- does a little bit of code cleanup