Track `GlobalRef` consistently
Currently dependencies for `GlobalRef` and `Symbol` are tracked
separately. Unfortunately, our lowering is not consistent about which
one it uses (after lowering, all `Symbols` refer to globals in the
current module), so the same variable can appear twice in the same
expression, once as a `GlobalRef`, once as a symbol. Currently, in that
situation, this package misses a dependency, because it considers them
two separate objects. Fix that by always normalizing to `GlobalRef`.
This does require the caller to pass through a Module, but the
alternative, would be for the caller to keep track of these dependencies
implicitly, which would be painful. I think consistently using
`GlobalRef` here would be cleanest.