adjustments to v1.12 (#124)
* adjustments to v1.12

There have been changes to type definitions and global variable
assignments to v1.12, and this commit tries to adjust the code base to
it. There might be some missing pieces still, this allow us to reduce
the failures in the test suite.

* fix up

* more fix

* rm `trackedheads`

On recent Julia versions we can just pattern match against `:method`
expressions.

* simplify `targetstore` creation

* Update src/codeedges.jl